pyproject.toml
[tool.poetry]
name = "pycdn"
version = "0.1.0"
description = ""
authors = ["Rafal <stapinskirafal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.81.0"
httpx = "^0.23.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
gunicorn = "^20.1.0"
uvloop = "^0.16.0"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pylint = "^2.15.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
known_local_folder = ["test"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.black]
line-length = 120

[tool.pylint.messages_control]
disable = "all"
enable = "unused-import"
